Hardening bandit in the face of buggy plugins or odd ASTs
When running in normal mode bandit will eat exceptions from a test
and continue onto the next one. This is a best effort attempt to
remain functional when we hit strange ASTs or plugin bugs. We also
try to log a useful message for bug reports.

In debug mode we still dump the info string but also re-raise the
error so developers can see what exploaded and why.

Change-Id: Ibb6aa489691c552cefe2f4309328a4725c5ec565