Further decorator changes and plugin migration
Continuing to deliver the refactor discussed during the OSSG mid-cycle.

This change completes the definition of new decorators to be applied
at the plugin/test level.  It also migrates existing tests over to use
one of the new decorators, using checks() instead of the checks_*()
functions.

With the checks_*() decorators we had tried to abstract away from the
AST nodes.  That was overly-complicating the framework, and there's
value to be had from understanding the AST, so we're removing that
abstraction.

Tests will now be able to use the checks() decorator and include the
AST nodes they can be executed against.  For example:

@checks('Import', 'ImportFrom')
def random_lib_imports(context):
[..]

@checks('Call')
def ssl_with_bad_version(context, config):
[..]

This change also cleans up the definition of additional new decorators
(severity(), confidence(), category(), title(), uuid()) which future
commits will build out the framework further to utilize.

Change-Id: I32017d21ed897b087ddfc249ebb5b604a9a832fa