Adds line ranges, DRYs code, fixes #nosec
This adds line ranges instead of just the line the warning starts on. Shuffle
some code around into methods so we can DRY up the code. Since we now have line
ranges, the #nosec flag can be found accross lines so it's no longer required
that #nosec be on the line that warning starts.

Change-Id: Id63cf0fa5d9882eec7c6d388054bc678fb6e0457