Tweaking severity for a few plugins
This patch sets exec down to a warning (instead of an error), so
that it is consistent with other possible code execution vectors
like eval.  This also removes a noisy info level message about
general subprocess calls, which will be moved to a separate
plugin soon.  Finally, this sets chmod group executable to a
warning instead of an error, because it is less dangerous since
groups can usually be trusted (somewhat).

Change-Id: I120876389d9395f6b007c051e0535a6b09ef7e63