Add error checks/handing around output file case
This changes partially addresses a reported bug by adding a number of
checks regarding the output file that can be specified by using -o on
the Bandit command line.

We check for cases where the file already exists, where the specified
destination is a directory, and where the specified destination is
not writable.  If any of these cases are identified, then we print an
error and exit before we begin executing any tests.

Change-Id: I9c5ba2f9973f8bf8b2456ece9862f6b701102e6a
Partial-Bug: #1432894