Fix a couple of issues with handling multi-line strings
In some cases, in multi-line statements, particularly in the case
of multi-line strings, some lines would be missing from the line
range. It is now implemented to use the range of all lines between
the first and last lines.

Also the code in the string node visitor which was used to ignore
docstrings was incomplete, in that docstrings can be seen as ast
expressions with ast string values.  The string node visitor now
excludes these expressions as well.

Change-Id: I1cd9fb83351f9fb3ee0b60778518d331bf0f66a1
Closes-Bug: #1432807