Fixing a bug with files listing when a file was skipped
This fixes a bug where if a file was skipped, it still showed in
the total listing of files, and so it could be associated with the
score of a non-skipped file. To fix these, when a file is skipped
we also remove it from the list of files in scope.

Change-Id: Ifdfe64b5720e1f809385140960b7b2c1d2238a92