fixing bandits config settings
This will now look for a 'bandit.yaml' file in the local folder
before falling back to ${HOME}/.config/bandit/bandit.yaml and
finally /etc/bandit/bandit.yaml. If -c is used to explicitly
specifiy a config file this will override all other behaviour.
If the specified file is not found then an error is reported.

Change-Id: Ie49a98f970ab31e9709e93ce06687f6457f03dce