Changing config file search paths
This commit changes the location that the Bandit config file,
bandit.yaml, is stored and how it is packaged.  Previously, the
config file was listed as a data_file which is supposed to be
outside of the Bandit installed package.  This meant that
depending on the system it might be installed in different places
(/etc, or /usr/lcoal/etc, for example).  When Bandit was installed
in a virutal environment the installed location would change once
again.  Another disadvantage to this approach is that installing
Bandit might require sudo, and Bandit might not clean up its
config properly.

This commit changes the packaging so that bandit.yaml is always
installed in bandit/config/bandit.yaml.  If there is a bandit
config file in the current directory or the user's home directory,
these are still preferred.

Change-Id: I5f971aa208dd2599f852b5253b4401990201cc8f