Adding /usr/local/etc/bandit/ to paths for Bandit config checking
When a user uses pip to install Bandit globally, the Bandit config
file is installed into /usr/local/etc/bandit, which is not one of
the paths which are checked during startup.  This leads to the
situation where a user runs "pip install bandit" and can't run
because they don't have a config file.  This commit fixes the
issue.

Change-Id: I5f971aa208dd2599f852b5253b4401990201cc8f