Added additional scope functions to StatementBuffer
I have made some changes to add additional scope awareness to the
StatementBuffer class. Additional functions have been added to allow
you to retrieve the lines of the parent of a given line, or all
children of a given line. Used in combination, it is possible to
retrieve all siblings of a given line.

StatementBuffer also now holds an archive of the AST of every statement
in it's unaltered state, which can be accessed using the new get_node
function, for example calling get_node(get_parent(line_no)) will
return the parent node of the given line, which includes the AST
of that node and all child nodes. Note, all the new functions take
a line number as a parameter, and return line numbers, or an AST
node for get_node().

Read the code for more info, or contact dave.belcher@hp.com for help

Change-Id: I2b0a6910fde06a8d02ea7878e1552719640302b2