Refactor BanditResultStore.report
Instead of adding a new elif branch inside that try-except block, it
makes more sense to move the actual writing of results to a private
method. That keeps the try portion of the try-except block small and
allows for new additions to the if/elif/else tree to be localized to
one, well-named method.

Change-Id: I81f570a6ac5c9c97f9fbedfc3fd34ebfe61c6e15