Add subunit output format
This commit adds support to bandit to generate a subunit stream
output reflecting the results of the run. It treats the subunit test
id as the filename with the check name  in '[]' appended to the end.
The default test runners (namely testr) used by other OpenStack
projects emit subunit streams. These are being used in a variety of
different ways so adding initial support to bandit to do the same is
a good start.

Change-Id: Idcb8463fe8b313bb2e4b8b96afc77afa7866eeb3