Removing statement buffer
The statement buffer previously mangaled the AST to produce chunks
that were passed, in isolation, to the node visitor. This cuased
some tests to become imposible, if the needed data from two chunks
had been seperated. This change fixes things by allowing the
node_visitor to process the AST directly.

To make this work it was necessary to embelish the AST with a few
extra bits of data, a node parent and sibling. This is a none
destructive change that does not limit any of the testing potential
of bandit comparied to the raw AST (indeed it enhances it).

As a beneficial side effect, "nosec" comments can now be used to
skip entire chunks of code, not just lines, for example

try:
 ...
except: #nosec
 ...

will now skip everything in except block.

Change-Id: I47c4963b143868a2bd2f171dbcf39e4f97a929da