Add tool for reporting Bandit OpenStack coverage
This script is intended for inclusion in the Bandit repository and usage
by the Bandit and other OpenStack teams.  It is not intended for inclusion
in Bandit packages / distribution.

It executes against Jenkins and Zuul configuration files in the
openstack-infra/project-config repository, parses out Bandit jobs
and tests, and prints a summary of results.

Co-Authored-By: Ian Cordasco <graffatcolmingov@gmail.com>
Change-Id: I2d133146223a6b185347662b47cc2bd6355a0900