Add a new check for weak RSA and DSA key sizes
This patch adds a new check to bandit that scans for key sizes of
RSA and DSA algorithms when using the pycrypto or cryptography.io
modules.

It also adds another method to the context to get the value of a
keyword argument.  This is needed because check_call_arg_value
only allows a comparision with known values and in this case the
values are variable (< 2048).

Change-Id: I7d740eccc73a49f7ee133d90177d19d5ef7b02ba