Add bytes support and fix hardcoded passwords
Add support for Bytes AST node which is only available in python 3. It
provides the same context object as Str nodes, but the type of .s can be
easily checked in the plugin itself.

Functions safe_unicode/safe_str were only valid in case of python 2.
They could not work as required in python 3, so they were removed.
Hardcoded passwords plugin was the only site where they were used.

Closes-bug: 1486310
Change-Id: I5df7f14b2b4985ed14386239e9e506d4a1549e42