Adding test tool for check OpenStack projects' Bandit job
This commit adds to the openstack coverage tool.  Specifically we
add a '-t' option.  If provided, this option will git clone any
project which uses Bandit in a job, run the tox Bandit job,
capture the output of any failed run, and display a summary table
at the end.

This tool is to support pre-release Bandit checking to make sure
that our changes haven't introduced any new issues in projects
which use it.

Change-Id: I321bcb15b59e3ee00ed2f2c6c2c890b77f30370e