Replace incorrect safe_str
The way safe_str is used is incorrect in case of python 3 and not
consistent in case of python 2. For example the following outputs depend
on the state of the whole string rather than each character in py2:

safe_str("\u0000") => "\x00"
safe_str("\uffff") => "\\uffff"
safe_str("\u0000\uffff") => "\\x00\\uffff"

(null is encoded once as null byte, once as string "\x00")

The replacement function applies to both Str and Bytes node contexts and
ensures that both representations always agree, at least in case of the
first 256 byte values / codepoints.

The result of Context.string_val_as_escaped_bytes is always printable
ascii and should be used when strings have unknown encoding, but should
be compared to a known value.

Closes-bug: 1491313
Change-Id: I7b065d22f03c10b9d3957a403c4439d2a7d74e78