Collecting metrics code in one place
The metrics gathering code was spread out throught the code base
(manager, node_visitor, reporters). This gathers as much as we
can into a single new class. Functionality is unchanged, coverage
is 100%

It also tweeks the interface to the node visitor. Process() now
takes a list of lines, rather than the file object. It would only
use this file object to read lines, this is now done in the
manager. Most metrics code is now called by the manager, the
exception being note_nosec() as this has to be called by the
node_visitor still.

Change-Id: Ibfae3c2b229fda999b998730d4c9301f4a7ccc45