Fixing Traceback with Bad File
This commit fixes a bug that was causing a traceback when running
Bandit against a single file with bad syntax.  The total metrics
count dictionary previously had no initial value, so if we had
a problem updating with new ones, we'd end up in a situation where
we never had a valid total count.  This fixes it by initializing
totals to 0 at metrics init time.

Change-Id: Ib7f1f54d0d6a0e9ffbb2cc3fc40b17b0c50ce0ca
Closes-Bug: #1512957
Closes-Bug: #1511431