Improving Bandit Baseline Reporting
This commit changes the Bandit Baseline reporting functionality.
The previous behavior was to compare each file and if the issue
total was different we'd output all of the issues from the file.

While this worked to implement a gate it wasn't easy to use for
penetration testing and the report generated could be difficult
for developers to parse to figure out why the gate broke.

The new approach will pair down the issues to determine exactly
which were not present in the baseline.  Next since code lines
can move aorund we can't figure out exactly where in the file
the issue is, but we can present a list of candidate issues.

A candidate issue matches the core properties of the finding
including file, severity, confidence, and issue text.

Since we're now presenting candidate issues for each finding we
also need a way to display that.  Baseline will now use separate
specifically designed formatters that show candidate issues.

Change-Id: Id5aa45ac4977e84f1c08a6b8e142acc6e5fc7a88