Adding the Text Baseline Formatter
This commit sets up using baseline formatters.  First we have to
check that the user has picked a support baseline output format
when they are using the baseline feature.  Assuming the user is
using a supported format and the baseline feature we'll
automatically pick the baseline version for them.

Baseline formatters are very similar to the normal version with
two changes.  First we don't output the code location for the
issue we found (since we don't know it).  Second, for each finding
we'll output a list of candidate findings with code location.  The
idea is that the user can pick out which of the candidates is the
actual new issue.

Change-Id: Ic91602ce7e9a970037113fec4b434c1554cf720b