Adding HTML baseline formatter
This commit adds the HTML baseline formatter which works the same
as the the normal HTML formatter except it displays candidate
issues in the case where Bandit can't determine which is the
matching issue.

It also spiffs up the HTML report a bit.  Similar enhancements
to normal HTML reporting are coming in a subsequent patch.

Change-Id: Ic4fa9b81cea094e62850a9fefffba8e966145695