Fixing a simple issue in results count to fix exit code
This commit fixes a simple mistake in the baseline code that was
causing a baseline run with no new issues to return a fail.

Basically we were deleting all objects from the baseline list, and
then using the length of the baseline list to check if we were in
baseline mode.  Since the length was now 0, we decided we weren't
running baseline.

Now we make a copy of the list and delete issues from the copy to
leave the original list alone.  Also we make the code more clear
by doing an explicit list of get_issue_list instead of length
of filter().

Change-Id: I32af0a1291cb94a527f31f4fcfe8faa7be302095