Adding bandit-baseline tool
This commit adds a tool which will run Bandit against the parent
commit of a current branch, and then run Bandit in baseline mode
using the parent's results as the baseline.  Any options that are
supplied to the script will be passed as options to Bandit (for
example severity filters, targets, etc).

By including this tool we can allow projects to run Bandit
baseline as part of their existing tox jobs.

Change-Id: Iaa1314aa348c7c5ca03c5c8b7dcfee456f279e56