Allow list of tests to skip to be specified on command line
As an alternative to defining a profile, it's very handy to have
a way to choose a list of tests to include straight from the CLI.

This patch maintains the existing -t option to pass a list of plugin
names or IDs, and adds a mutually exclusive -s option to pass a list
of plugin names or IDs to skip.

When using -s or --skip, Bandit will run all plugins, except those
on the list specified on the command line.

Change-Id: Icd360eea89dd805987332f638166a1e09edee641