Breaking out blacklists
This change removes the old blacklist plugins and replaces them
with new built in functionality that loads blacklist item data
from a new plugin entry point. The new test also improve on the
old functionality that was broken in the following way:

import xml.sax # issue found OK
from xml import sax # no issue found, wrong

Finally, this patch removes the use of filename style wild cards
such as * from the import blacklist matching, as this was not being
used. Both this test and the old ones will alert on any import from
within the blacklisted namespace.

Note: this change also requires the Issue object to take an
explicit bandit id, as we can't get it from the test func in the
case of blacklists.

Change-Id: I98af6daf3c54561c0e4b399605ea615b42b7b283