This permits blacklist data to be filtered by ID
Note that we now prefer IDs over names when giving include/exclude
lists. If we are given a list with plugin names in it we will
convert them to IDs. This means we can also have mixed lists.
However, the builtin tests and the plugin data have no names, thus
they can only be filtered by ID.

Most of the time filtering data should be given as IDs the name
converting logic exists only for backwards compatibility with old
configs.

Change-Id: I74d80e99d9705a4bd5e786fef3e1ec68f3c4e344