Fixing hang in get_module_qualname_from_path
This commit fixes a possible hang in the get_module_qualname...
function.  When running against a '.' target, the while statement
would loop forever looking for a head of '/', when all it would
ever find was '.'.  Now either option is enough to break the loop.

A head of '.' already indicates we're as far down as we can get
since there is no non-file path left.

Closes-Bug: #1538633
Change-Id: I4aca741b816e4203d6b4da4e62c5edd37d553da4