Use sys.exit(args) instead of logger for errors
Passing the error string to sys.exit is preferred to using a
logger to output error messages. The exit codes themselves will
be 1 for any non-None argument. Some of the 2 exit codes as a
result will now be 1, but there is no standard set of exit codes
as indicate in the following link.

https://docs.python.org/2/library/sys.html

Change-Id: Iea43a20075f7edf90d724602d6745e37c89e11a3