Adding support for new style profile
This commit adds support for the new style of Bandit profile.
These are typically generated by the Bandit Config Generator tool
and may be used to select or exclude tests or override the
settings for plugins.

Since we still need support for legacy config/profiles and it
doesn't make sense to add a new flag this overloads the profile
setting.  It will first detect if the profile is a file or a
legacy profile.  This is written with the idea that it will be
easy to find and get rid of legacy code when we're ready.

Change-Id: I939639be795249eee9d3598b4e9f190e64efe405