Breaking up blacklist import IDs
This patch breaks up the blacklist imports sets to be one import
per ID (with the exception of things like pickle/cPickle etc).
This is done to give us more control over filtering stuff. It is
important to make any changes to IDs before 1.0 release as they
will form part of the 1.0 API contract (this renumbers some XML
IDs).

Change-Id: Ib575f6155a58f1795fe2c05fabdba4dbf1a89519