Improving config handling
This patch adds a validation step to the config class that can
be used to detect bad configs. For now it just asserts that if
legacy blacklist tests are mentioned the config contains the
required data block for them.

Additionally, this also removes various places in our test set
where a config file is specified. This was only done to satisfy
the old behaviour when a config was not optional.

Finally, this detects when a config has legacy data in it and
prints a deprecation warning.

Test coverage of the config class is now raised to 100%

Change-Id: I492a20f9b9f421d32e3e72eaa15f88c34c3d11e8