Catch general exception on per-file basis
This modifies the Bandit manager to catch a general Exception on a per-file
basis. Then an exception does occur, the name of the file is emitted and
the file is logged as a 'skipped file' for inclusion in the end-of-run
output.  When run in debug mode, a traceback will also be printed.

The change also adds a new test targeting this case, along with a new
example file (nonsense2.py is gzipped nonsense.py) to trigger the test.

Change-Id: I86e648890dddcc5c2fff7dd9844678e990b0cd63
Closes-Bug: #1498258